Fixed bug, version 0.1.1711

+++ bare/www/core/js/services/geolocation.factory.js
@@ -1104,8 +1104,10 @@ angular.module('webmapp')
                     return BackgroundGeolocation.removeAllListeners(event);
                 });
 
-                turnOffRotationAndFollow();
-                MapService.removePosition();
+                if (MapService.hasMap()) {
+                    turnOffRotationAndFollow();
+                    MapService.removePosition();
+                }
                 BackgroundGeolocation.stop();
                 state.reset();
                 recordingState.reset();
